sdm630.dtsi : increase xo_clock to 21.5Mhz
xo_clock is responsible for generating a stable and extremely reliable clock for the rest of the system, so many components use it as the basis for their multipliers in Snapdragon processors. The most common are: CPU, GPU, memory controller (ram), wireless controller, audio encoder.

Standard clock for SDM630 = 19.2MHz

By making a small adjustment to this clock, we can increase/decrease the frequency of these components without having to interfere directly with their code.

for a light and safe overclock, we will increase the frequency to 21.2MHz to maintain stability without having to interfere with the component voltages.

by Chanz22